+++ backend/server.js
@@ -5,7 +5,7 @@ const cors = require("cors");
 const app = express();
 
 var corsOptions = {
-    origin: "http://localhost:2999"
+    origin: "http://localhost:8081"
 };
 
 app.use(cors(corsOptions));
@@ -15,6 +15,7 @@ app.use(bodyParser.json());
 
 // parse requests of content-type - application/x-www-form-urlencoded
 app.use(bodyParser.urlencoded({ extended: true }));
+
 
 const db = require("./app/models");
 db.mongoose
@@ -30,13 +31,17 @@ db.mongoose
         process.exit();
     });
 
+
+
 // simple route
 app.get("/", (req, res) => {
-    res.json({ message: "Bienvene sur notre application de SIGB." });
+    res.json({ message: "Bienvenue sur notre application de SIGB." });
 });
 
+require("./app/routes/book.routes")(app);
+
 // set port, listen for requests
-const PORT = process.env.PORT || 3000;
+const PORT = process.env.PORT || 8080;
 app.listen(PORT, () => {
     console.log(`Server is running on port ${PORT}.`);
 });